+++ frontend/src/pages/error.tsx
@@ -4,7 +4,6 @@ import Meta from "../components/Meta";
 import useTranslate from "../hooks/useTranslate.hook";
 import { useRouter } from "next/router";
 import { FormattedMessage } from "react-intl";
-import { safeRedirectPath } from "../utils/router.util";
 
 const useStyle = createStyles({
   title: {
@@ -38,13 +37,8 @@ export default function Error() {
             )}
           />
         </Text>
-        <Button
-          mt="xl"
-          onClick={() =>
-            router.push(safeRedirectPath(router.query.redirect as string))
-          }
-        >
-          {t("error.button.back")}
+        <Button mt="xl" onClick={() => router.push("/")}>
+          {"Go Home"}
         </Button>
       </Stack>
     </>